+++ internal/compiler/expression_tree.rs
@@ -40,7 +40,7 @@ pub enum BuiltinFunction {
     SetFocusItem,
     ShowPopupWindow,
     ClosePopupWindow,
-    SelectRange,
+    SetSelectionOffsets,
     /// A function that belongs to an item (such as TextInput's select-all function).
     ItemMemberFunction(String),
     /// the "42".to_float()
@@ -135,7 +135,7 @@ impl BuiltinFunction {
                     args: vec![Type::ElementReference],
                 }
             }
-            BuiltinFunction::SelectRange => Type::Function {
+            BuiltinFunction::SetSelectionOffsets => Type::Function {
                 return_type: Box::new(Type::Void),
                 args: vec![Type::ElementReference, Type::Int32, Type::Int32],
             },
@@ -253,7 +253,7 @@ impl BuiltinFunction {
             | BuiltinFunction::ATan => true,
             BuiltinFunction::SetFocusItem => false,
             BuiltinFunction::ShowPopupWindow | BuiltinFunction::ClosePopupWindow => false,
-            BuiltinFunction::SelectRange => false,
+            BuiltinFunction::SetSelectionOffsets => false,
             BuiltinFunction::ItemMemberFunction(..) => false,
             BuiltinFunction::StringToFloat | BuiltinFunction::StringIsFloat => true,
             BuiltinFunction::ColorBrighter
@@ -307,7 +307,7 @@ impl BuiltinFunction {
             | BuiltinFunction::ATan => true,
             BuiltinFunction::SetFocusItem => false,
             BuiltinFunction::ShowPopupWindow | BuiltinFunction::ClosePopupWindow => false,
-            BuiltinFunction::SelectRange => false,
+            BuiltinFunction::SetSelectionOffsets => false,
             BuiltinFunction::ItemMemberFunction(..) => false,
             BuiltinFunction::StringToFloat | BuiltinFunction::StringIsFloat => true,
             BuiltinFunction::ColorBrighter