+++ internal/compiler/generator/rust.rs
@@ -2365,7 +2365,7 @@ fn compile_builtin_function_call(
                 sp::WindowInner::from_pub(#window_adapter_tokens.window()).close_popup()
             )
         }
-        BuiltinFunction::SelectRange => {
+        BuiltinFunction::SetSelectionOffsets => {
             if let [llr::Expression::PropertyReference(pr), from, to] = arguments {
                 let item = access_member(pr, ctx);
                 let item_rc = access_item_rc(pr, ctx);
@@ -2374,10 +2374,10 @@ fn compile_builtin_function_call(
                 let to = compile_expression(to, ctx);
 
                 quote!(
-                    #item.select(#window_adapter_tokens, #item_rc, #from as i32, #to as i32)
+                    #item.set_selection_offsets(#window_adapter_tokens, #item_rc, #from as i32, #to as i32)
                 )
             } else {
-                panic!("internal error: invalid args to TextInputSelect {:?}", arguments)
+                panic!("internal error: invalid args to set-selection-offsets {:?}", arguments)
             }
         }
         BuiltinFunction::ItemMemberFunction(name) => {