+++ internal/compiler/typeregister.rs
@@ -339,10 +339,10 @@ impl TypeRegister {
             ElementType::Builtin(ref mut b) => {
                 let text_input = Rc::get_mut(b).unwrap();
                 text_input.properties.insert(
-                    "select".into(),
-                    BuiltinPropertyInfo::new(BuiltinFunction::SelectRange.ty()),
+                    "set-selection-offsets".into(),
+                    BuiltinPropertyInfo::new(BuiltinFunction::SetSelectionOffsets.ty()),
                 );
-                text_input.member_functions.insert("select".into(), BuiltinFunction::SelectRange);
+                text_input.member_functions.insert("set-selection-offsets".into(), BuiltinFunction::SetSelectionOffsets);
             }
 
             _ => unreachable!(),